requirements.txt
astroid==2.13.5
async-generator==1.10
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
dill==0.3.6
EasyProcess==1.1
exceptiongroup==1.1.1
h11==0.13.0
html-testRunner==1.2.1
idna==3.3
iniconfig==1.1.1
install==1.3.5
isort==5.12.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
outcome==1.2.0
packaging==21.3
parameterized==0.8.1
Pillow==9.2.0
platformdirs==3.2.0
pluggy==1.0.0
psycopg2==2.9.5
py==1.11.0
pycparser==2.21
pylint==2.15.3
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-dependency==0.5.1
pytest-html==3.1.1
pytest-html-reporter==0.2.9
pytest-metadata==2.0.2
pytest-ordering==0.6
python-dateutil==2.8.2
python-dotenv==0.20.0
PyVirtualDisplay==0.2.4
requests==2.28.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
tomlkit==0.11.7
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
webdriver-manager==3.8.3
wrapt==1.15.0
wsproto==1.1.0
